Add more globals and add doc
